Fix tests and lint
